project3_traffic_prediction/requirements-visualization.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow>=10.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
gunicorn>=21.2.0
psutil>=5.9.0
